requirements.txt
simple-salesforce
python-dotenv
pytest
requests
